refactor: extract crate::math from skel/skinning
