+++ src/schemas/skel/skinning.rs
@@ -1,137 +1,14 @@
-//! Pure-math skinning helpers — linear blend skinning, transform
-//! composition, blend-shape application.
+//! Linear blend skinning + blend-shape application — the skinning
+//! pipeline math that consumes pre-resolved joint transforms.
 //!
-//! All matrix data uses USD's textual `matrix4d` layout: a flat
-//! `[f64; 16]` storing the matrix in row-major order (the same order
-//! the USDA writer emits). All operations live in this module so the
-//! `skel` crate doesn't grow a top-level math dependency — callers
-//! can keep using their own math library and only convert at the
-//! interface.
+//! All matrix data uses USD's textual `matrix4d` layout (row-major
+//! flat `[f64; 16]`). The 4×4 algebra itself lives in [`crate::math`];
+//! this module imports those helpers and adds the skinning-specific
+//! transforms on top.
+
+use crate::math::{mat4_inverse, mat4_mul, mat4_transform_point, mat4_transform_vec, IDENTITY_MAT4};
 
 use super::topology::{Topology, NO_PARENT};
-
-/// 4×4 identity in USD's `matrix4d` row-major layout.
-pub const IDENTITY_MAT4: [f64; 16] = [
-    1.0, 0.0, 0.0, 0.0, //
-    0.0, 1.0, 0.0, 0.0, //
-    0.0, 0.0, 1.0, 0.0, //
-    0.0, 0.0, 0.0, 1.0, //
-];
-
-/// `out = a · b` for two 4×4 row-major matrices.
-///
-/// USD authors transforms in row-major form and applies them by
-/// pre-multiplying row vectors: `v_out = v_in · M`. Sequences of
-/// transforms therefore chain left-to-right: to place a child's local
-/// transform under its parent's, compose `(M_local · M_parent)` — the
-/// same form the UsdSkel spec writes as
-/// `jointLocalSpaceTransform * parentJointSkelSpaceTransform`.
-pub fn mat4_mul(a: &[f64; 16], b: &[f64; 16]) -> [f64; 16] {
-    let mut out = [0.0f64; 16];
-    for r in 0..4 {
-        for c in 0..4 {
-            let mut v = 0.0;
-            for k in 0..4 {
-                v += a[r * 4 + k] * b[k * 4 + c];
-            }
-            out[r * 4 + c] = v;
-        }
-    }
-    out
-}
-
-/// Apply a 4×4 row-major matrix to a position. Equivalent to
-/// `(p.x, p.y, p.z, 1) · M` and then dropping `w`.
-pub fn mat4_transform_point(m: &[f64; 16], p: [f32; 3]) -> [f32; 3] {
-    let x = p[0] as f64;
-    let y = p[1] as f64;
-    let z = p[2] as f64;
-    let nx = x * m[0] + y * m[4] + z * m[8] + m[12];
-    let ny = x * m[1] + y * m[5] + z * m[9] + m[13];
-    let nz = x * m[2] + y * m[6] + z * m[10] + m[14];
-    [nx as f32, ny as f32, nz as f32]
-}
-
-/// Apply the rotational + scale part of a matrix to a direction
-/// vector (e.g. a normal). Ignores the translation row.
-pub fn mat4_transform_vec(m: &[f64; 16], v: [f32; 3]) -> [f32; 3] {
-    let x = v[0] as f64;
-    let y = v[1] as f64;
-    let z = v[2] as f64;
-    let nx = x * m[0] + y * m[4] + z * m[8];
-    let ny = x * m[1] + y * m[5] + z * m[9];
-    let nz = x * m[2] + y * m[6] + z * m[10];
-    [nx as f32, ny as f32, nz as f32]
-}
-
-/// Singularity threshold for [`mat4_inverse`]. `f64::EPSILON` (≈ 2.2e-16)
-/// is tight enough that ill-conditioned real-world bind matrices slip
-/// through and produce numerically-unstable inverses; `1e-12` matches
-/// the tolerance Pixar's reference inverter uses for skel data.
-const MAT4_SINGULAR_DETERMINANT: f64 = 1e-12;
-
-/// Invert a 4×4 row-major matrix using cofactor expansion.
-/// Returns [`None`] when the matrix is singular or near-singular
-/// (`|det| < `[`MAT4_SINGULAR_DETERMINANT`]).
-pub fn mat4_inverse(m: &[f64; 16]) -> Option<[f64; 16]> {
-    let mut inv = [0.0f64; 16];
-
-    inv[0] =
-        m[5] * m[10] * m[15] - m[5] * m[11] * m[14] - m[9] * m[6] * m[15] + m[9] * m[7] * m[14] + m[13] * m[6] * m[11]
-            - m[13] * m[7] * m[10];
-    inv[4] =
-        -m[4] * m[10] * m[15] + m[4] * m[11] * m[14] + m[8] * m[6] * m[15] - m[8] * m[7] * m[14] - m[12] * m[6] * m[11]
-            + m[12] * m[7] * m[10];
-    inv[8] =
-        m[4] * m[9] * m[15] - m[4] * m[11] * m[13] - m[8] * m[5] * m[15] + m[8] * m[7] * m[13] + m[12] * m[5] * m[11]
-            - m[12] * m[7] * m[9];
-    inv[12] =
-        -m[4] * m[9] * m[14] + m[4] * m[10] * m[13] + m[8] * m[5] * m[14] - m[8] * m[6] * m[13] - m[12] * m[5] * m[10]
-            + m[12] * m[6] * m[9];
-    inv[1] =
-        -m[1] * m[10] * m[15] + m[1] * m[11] * m[14] + m[9] * m[2] * m[15] - m[9] * m[3] * m[14] - m[13] * m[2] * m[11]
-            + m[13] * m[3] * m[10];
-    inv[5] =
-        m[0] * m[10] * m[15] - m[0] * m[11] * m[14] - m[8] * m[2] * m[15] + m[8] * m[3] * m[14] + m[12] * m[2] * m[11]
-            - m[12] * m[3] * m[10];
-    inv[9] =
-        -m[0] * m[9] * m[15] + m[0] * m[11] * m[13] + m[8] * m[1] * m[15] - m[8] * m[3] * m[13] - m[12] * m[1] * m[11]
-            + m[12] * m[3] * m[9];
-    inv[13] =
-        m[0] * m[9] * m[14] - m[0] * m[10] * m[13] - m[8] * m[1] * m[14] + m[8] * m[2] * m[13] + m[12] * m[1] * m[10]
-            - m[12] * m[2] * m[9];
-    inv[2] =
-        m[1] * m[6] * m[15] - m[1] * m[7] * m[14] - m[5] * m[2] * m[15] + m[5] * m[3] * m[14] + m[13] * m[2] * m[7]
-            - m[13] * m[3] * m[6];
-    inv[6] =
-        -m[0] * m[6] * m[15] + m[0] * m[7] * m[14] + m[4] * m[2] * m[15] - m[4] * m[3] * m[14] - m[12] * m[2] * m[7]
-            + m[12] * m[3] * m[6];
-    inv[10] =
-        m[0] * m[5] * m[15] - m[0] * m[7] * m[13] - m[4] * m[1] * m[15] + m[4] * m[3] * m[13] + m[12] * m[1] * m[7]
-            - m[12] * m[3] * m[5];
-    inv[14] =
-        -m[0] * m[5] * m[14] + m[0] * m[6] * m[13] + m[4] * m[1] * m[14] - m[4] * m[2] * m[13] - m[12] * m[1] * m[6]
-            + m[12] * m[2] * m[5];
-    inv[3] =
-        -m[1] * m[6] * m[11] + m[1] * m[7] * m[10] + m[5] * m[2] * m[11] - m[5] * m[3] * m[10] - m[9] * m[2] * m[7]
-            + m[9] * m[3] * m[6];
-    inv[7] = m[0] * m[6] * m[11] - m[0] * m[7] * m[10] - m[4] * m[2] * m[11] + m[4] * m[3] * m[10] + m[8] * m[2] * m[7]
-        - m[8] * m[3] * m[6];
-    inv[11] = -m[0] * m[5] * m[11] + m[0] * m[7] * m[9] + m[4] * m[1] * m[11] - m[4] * m[3] * m[9] - m[8] * m[1] * m[7]
-        + m[8] * m[3] * m[5];
-    inv[15] = m[0] * m[5] * m[10] - m[0] * m[6] * m[9] - m[4] * m[1] * m[10] + m[4] * m[2] * m[9] + m[8] * m[1] * m[6]
-        - m[8] * m[2] * m[5];
-
-    let det = m[0] * inv[0] + m[1] * inv[4] + m[2] * inv[8] + m[3] * inv[12];
-    if det.abs() < MAT4_SINGULAR_DETERMINANT {
-        return None;
-    }
-    let inv_det = 1.0 / det;
-    for v in &mut inv {
-        *v *= inv_det;
-    }
-    Some(inv)
-}
 
 /// Compose joint-local transforms into skeleton-space transforms by
 /// concatenating each joint's local transform with its parent's
@@ -495,106 +372,6 @@ mod tests {
         m[13] = y;
         m[14] = z;
         m
-    }
-
-    #[test]
-    fn mat4_mul_identity_is_noop() {
-        let m = translation(1.0, 2.0, 3.0);
-        assert_eq!(mat4_mul(&IDENTITY_MAT4, &m), m);
-        assert_eq!(mat4_mul(&m, &IDENTITY_MAT4), m);
-    }
-
-    #[test]
-    fn point_transform_applies_translation() {
-        let m = translation(10.0, 20.0, 30.0);
-        let p = mat4_transform_point(&m, [1.0, 2.0, 3.0]);
-        assert_eq!(p, [11.0, 22.0, 33.0]);
-    }
-
-    #[test]
-    fn vec_transform_ignores_translation() {
-        let m = translation(10.0, 20.0, 30.0);
-        let v = mat4_transform_vec(&m, [1.0, 2.0, 3.0]);
-        assert_eq!(v, [1.0, 2.0, 3.0]);
-    }
-
-    #[test]
-    fn inverse_round_trips() {
-        let m = translation(5.0, -2.0, 7.0);
-        let inv = mat4_inverse(&m).unwrap();
-        let id = mat4_mul(&m, &inv);
-        for i in 0..16 {
-            assert!((id[i] - IDENTITY_MAT4[i]).abs() < 1e-10, "row-col {i}: {id:?}");
-        }
-    }
-
-    #[test]
-    fn inverse_round_trips_with_rotation() {
-        // Build a non-trivial transform that exercises every off-diagonal
-        // term of the 3x3 sub-block plus the translation row. A translation-
-        // only round-trip is too easy — its inverse is just the negated row
-        // — and would not catch an `m[r*4 + c]` ↔ `m[c*4 + r]` typo in any
-        // of the 16 cofactor formulas in mat4_inverse.
-        let scale = {
-            let mut m = IDENTITY_MAT4;
-            m[0] = 2.0;
-            m[5] = 3.0;
-            m[10] = 0.5;
-            m
-        };
-        // 30° rotation around Z (row-vector form: v_new = v · R_z).
-        let rot_z = {
-            let c = 0.8660254037844387_f64;
-            let s = 0.5_f64;
-            let mut m = IDENTITY_MAT4;
-            m[0] = c;
-            m[1] = s;
-            m[4] = -s;
-            m[5] = c;
-            m
-        };
-        let trans = translation(7.0, -3.0, 11.0);
-
-        // M = scale · rot_z · trans. Order doesn't matter for the test;
-        // any invertible non-axis-aligned composition lights up the cofactors.
-        let m = mat4_mul(&mat4_mul(&scale, &rot_z), &trans);
-        let inv = mat4_inverse(&m).expect("non-singular");
-
-        // Both M · inv and inv · M must equal identity. Checking both
-        // catches any cofactor-expansion asymmetry that a one-sided check
-        // would miss.
-        let id1 = mat4_mul(&m, &inv);
-        let id2 = mat4_mul(&inv, &m);
-        for i in 0..16 {
-            assert!((id1[i] - IDENTITY_MAT4[i]).abs() < 1e-10, "M·inv off at {i}: {id1:?}");
-            assert!((id2[i] - IDENTITY_MAT4[i]).abs() < 1e-10, "inv·M off at {i}: {id2:?}");
-        }
-
-        // Round-trip a non-trivial point through M then inv — independent
-        // sanity check on top of the matrix identity.
-        let p = [4.0_f32, -2.5, 7.5];
-        let warped = mat4_transform_point(&m, p);
-        let back = mat4_transform_point(&inv, warped);
-        for k in 0..3 {
-            assert!((back[k] - p[k]).abs() < 1e-4, "point round-trip off: {back:?} vs {p:?}");
-        }
-    }
-
-    #[test]
-    fn inverse_rejects_near_singular() {
-        // A 4x4 with two identical rows is singular (det = 0); a
-        // matrix scaled by ~1e-7 across the board is well above
-        // f64::EPSILON but well below the 1e-12 skel threshold.
-        let mut near_singular = [0.0f64; 16];
-        // Diagonal scale of 1e-4 means det = 1e-16 — under 1e-12.
-        for i in 0..4 {
-            near_singular[i * 4 + i] = 1e-4;
-        }
-        assert!(mat4_inverse(&near_singular).is_none(), "should reject near-singular");
-
-        // Plain singular (row of zeros).
-        let singular = [0.0f64; 16];
-        assert!(mat4_inverse(&singular).is_none(), "should reject zero matrix");
     }
 
     #[test]
